refactor: move DeactivateTimeOutJobs outside JobTimeoutChecker to make it testable
(cherry picked from commit 73c737da568a3e945f0beba88b7095d5340f0efa)
