+++ engine/src/main/java/io/camunda/zeebe/engine/processing/job/DeactivateTimeOutJobs.java
@@ -1,0 +1,97 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Zeebe Community License 1.1. You may not use this file
+ * except in compliance with the Zeebe Community License 1.1.
+ */
+package io.camunda.zeebe.engine.processing.job;
+
+import static io.camunda.zeebe.scheduler.clock.ActorClock.currentTimeMillis;
+
+import io.camunda.zeebe.engine.state.immutable.JobState;
+import io.camunda.zeebe.engine.state.immutable.JobState.DeadlineIndex;
+import io.camunda.zeebe.protocol.record.intent.JobIntent;
+import io.camunda.zeebe.stream.api.ReadonlyStreamProcessorContext;
+import io.camunda.zeebe.stream.api.scheduling.Task;
+import io.camunda.zeebe.stream.api.scheduling.TaskResult;
+import io.camunda.zeebe.stream.api.scheduling.TaskResultBuilder;
+import java.time.Duration;
+import org.agrona.collections.MutableInteger;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+final class DeactivateTimeOutJobs implements Task {
+  private static final Logger LOG = LoggerFactory.getLogger(DeactivateTimeOutJobs.class);
+
+  private boolean shouldReschedule = false;
+
+  /** Keeps track of the timestamp to compare the message deadlines against. */
+  private long executionTimestamp = -1;
+
+  /** Keeps track of where to continue between iterations. */
+  private DeadlineIndex startAtIndex = null;
+
+  private final JobState state;
+  private ReadonlyStreamProcessorContext processingContext;
+  private final Duration pollingInterval;
+  private final int batchLimit;
+
+  public DeactivateTimeOutJobs(
+      final JobState state, final Duration pollingInterval, final int batchLimit) {
+    this.state = state;
+    this.pollingInterval = pollingInterval;
+    this.batchLimit = batchLimit;
+  }
+
+  public void schedule(final Duration idleInterval) {
+    if (shouldReschedule) {
+      processingContext.getScheduleService().runDelayed(idleInterval, this);
+    }
+  }
+
+  @Override
+  public TaskResult execute(final TaskResultBuilder taskResultBuilder) {
+    LOG.trace("Job timout checker running...");
+    if (executionTimestamp == -1) {
+      executionTimestamp = currentTimeMillis();
+    }
+
+    final var counter = new MutableInteger(0);
+
+    final DeadlineIndex nextIndex =
+        state.forEachTimedOutEntry(
+            executionTimestamp,
+            startAtIndex,
+            (key, record) -> {
+              if (counter.getAndIncrement() >= batchLimit) {
+                return false;
+              }
+
+              return taskResultBuilder.appendCommandRecord(key, JobIntent.TIME_OUT, record);
+            });
+
+    if (nextIndex != null) {
+      LOG.trace(
+          "Job timout checker yielded early. will reschedule immediately from where it left of.");
+      startAtIndex = nextIndex;
+      schedule(Duration.ZERO);
+    } else {
+      executionTimestamp = -1;
+      startAtIndex = null;
+      schedule(pollingInterval);
+    }
+
+    LOG.trace("{} jobs has been marked to timeout", counter.get());
+
+    return taskResultBuilder.build();
+  }
+
+  public void setProcessingContext(final ReadonlyStreamProcessorContext processingContext) {
+    this.processingContext = processingContext;
+  }
+
+  public void setShouldReschedule(final boolean shouldReschedule) {
+    this.shouldReschedule = shouldReschedule;
+  }
+}
